#423: setup 메서드 위치 이동
- EditToDoViewController에는 VIP 관련 로직만 두도록 setup 메서드의 위치를 이동하였습니다.
